rename gcpmanager to sci-gcp in code
