+++ config/sci-gcp/kustomization.yaml
@@ -1,8 +1,8 @@
 resources:
-- gcp-manager.yaml
+- ../sci
 apiVersion: kustomize.config.k8s.io/v1beta1
 kind: Kustomization
 images:
-- name: gcp-manager
-  newName: docker.io/substratusai/gcp-manager
+- name: sci
+  newName: docker.io/substratusai/sci-gcp
   newTag: v0.8.1
